Move helper functions and refactor tests
This commit relocates the helper functions from 'tests/middleware' to 'tests', adjusting import paths in various test files accordingly. Additionally, the 'client' fixture in 'test_endpoints' is replaced with a new 'client_with_mock_db' fixture, and a new 'client_with_db' fixture is introduced in 'fixtures.py'. This enhances the organization and maintainability of the test infrastructure.
